+++ src/lib/scrub.ts
@@ -16,11 +16,15 @@ export function scrubString(s: string): string {
 }
 
 // Harder redaction for text that leaves the device (support inbox): drop
-// every URL wholesale and erase every occurrence of the user's server host,
-// which also catches bare `host=…` fragments and log lines.
-export function redactHostAndUrls(text: string, host?: string): string {
+// every URL wholesale, erase every known server host (bare `host=…` fragments
+// and log lines carry hosts without a scheme, which the URL regex misses),
+// and blank bare IPv4 addresses as a catch-all for hosts we never parsed.
+export function redactHostAndUrls(text: string, hosts?: Array<string | undefined>): string {
   let out = text.replace(/https?:\/\/[^\s)\]}"']+/gi, "<redacted-url>")
-  if (host) out = out.split(host).join("<redacted-host>")
+  for (const host of hosts ?? []) {
+    if (host) out = out.split(host).join("<redacted-host>")
+  }
+  out = out.replace(/\b(?:\d{1,3}\.){3}\d{1,3}\b/g, "<redacted-ip>")
   return out
 }
